[RN] Add InfoDialogButton
