+++ react/features/invite/components/index.js
@@ -2,6 +2,5 @@
 
 export * from './add-people-dialog';
 export { DialInSummary } from './dial-in-summary';
-export { default as InfoDialogButton } from './InfoDialogButton';
-export { default as InviteButton } from './InviteButton';
+export * from './info-dialog';
 export * from './callee-info';
