--- react/features/invite/components/info-dialog/index.js
@@ -1,1 +1,0 @@
-export { default as InfoDialog } from './InfoDialog';
