+++ react/features/invite/components/info-dialog/index.web.js
@@ -1,0 +1,3 @@
+// @flow
+
+export * from './web';
