+++ react/features/invite/components/info-dialog/native/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as InfoDialogButton } from './InfoDialogButton';
