+++ react/features/invite/components/info-dialog/web/DialInNumber.js
@@ -1,8 +1,8 @@
-/* @flow */
+// @flow
 
 import React, { Component } from 'react';
 
-import { translate } from '../../../base/i18n';
+import { translate } from '../../../../base/i18n';
 
 /**
  * The type of the React {@code Component} props of {@link DialInNumber}.
