+++ react/features/invite/components/info-dialog/web/InfoDialogButton.js
@@ -4,18 +4,18 @@ import InlineDialog from '@atlaskit/inline-dialog';
 import React, { Component } from 'react';
 import type { Dispatch } from 'redux';
 
-import { createToolbarEvent, sendAnalytics } from '../../analytics';
-import { openDialog } from '../../base/dialog';
-import { translate } from '../../base/i18n';
-import { JitsiRecordingConstants } from '../../base/lib-jitsi-meet';
-import { getParticipantCount } from '../../base/participants';
-import { OverflowMenuItem } from '../../base/toolbox';
-import { connect } from '../../base/redux';
-import { getActiveSession } from '../../recording';
-import { ToolbarButton } from '../../toolbox';
-import { updateDialInNumbers } from '../actions';
+import { createToolbarEvent, sendAnalytics } from '../../../../analytics';
+import { openDialog } from '../../../../base/dialog';
+import { translate } from '../../../../base/i18n';
+import { JitsiRecordingConstants } from '../../../../base/lib-jitsi-meet';
+import { getParticipantCount } from '../../../../base/participants';
+import { OverflowMenuItem } from '../../../../base/toolbox';
+import { connect } from '../../../../base/redux';
+import { getActiveSession } from '../../../../recording';
+import { ToolbarButton } from '../../../../toolbox';
+import { updateDialInNumbers } from '../../../actions';
 
-import { InfoDialog } from './info-dialog';
+import InfoDialog from './InfoDialog';
 
 /**
  * The type of the React {@code Component} props of {@link InfoDialogButton}.
