+++ react/features/invite/components/info-dialog/web/PasswordForm.js
@@ -1,9 +1,9 @@
-/* @flow */
+// @flow
 
 import React, { Component } from 'react';
 
-import { translate } from '../../../base/i18n';
-import { LOCKED_LOCALLY } from '../../../room-lock';
+import { translate } from '../../../../base/i18n';
+import { LOCKED_LOCALLY } from '../../../../room-lock';
 
 /**
  * The type of the React {@code Component} props of {@link PasswordForm}.
