+++ react/features/invite/components/info-dialog/web/index.js
@@ -1,0 +1,4 @@
+// @flow
+
+export { default as InfoDialog } from './InfoDialog';
+export { default as InfoDialogButton } from './InfoDialogButton';
