+++ react/features/toolbox/components/native/OverflowMenu.js
@@ -10,7 +10,7 @@ import {
 } from '../../../base/dialog';
 import { connect } from '../../../base/redux';
 import { StyleType } from '../../../base/styles';
-import { InviteButton } from '../../../invite';
+import { InfoDialogButton, InviteButton } from '../../../invite';
 import { AudioRouteButton } from '../../../mobile/audio-mode';
 import { LiveStreamButton, RecordButton } from '../../../recording';
 import { RoomLockButton } from '../../../room-lock';
@@ -96,6 +96,7 @@ class OverflowMenu extends Component<Props> {
                 <LiveStreamButton { ...buttonProps } />
                 <TileViewButton { ...buttonProps } />
                 <InviteButton { ...buttonProps } />
+                <InfoDialogButton { ...buttonProps } />
                 <RaiseHandButton { ...buttonProps } />
             </BottomSheet>
         );
